+++ CHANGELOG.md
@@ -5,16 +5,26 @@ All notable changes to this project will be documented in this file. The format 
 ## [1.0.0] - 22-08-2025
 ## [1.0.1] - 07-11-2025
 ## [1.0.2] - 14-11-2025
+## [1.0.3] - 30-01-2026
+
 
 This is the first official release of _HelloID-Conn-Prov-Target-Caseware_. This release is based on template version _v3.1.0_.
 
 ### Added
 Enable.ps1
+    Enable.ps1 to enable Caseware accounts
 Disable.ps1
-Account Create
+    Disable.ps1 to enable Caseware accounts
+Permissions roles
+    To manage roles
+Permissions groups
+    To manage groups
+
 
 ### Changed
-Permissions will manage the groups instead of Roles.
+fieldMapping.json
+    Added extra fields enable and disable Caseware accounts. 
 ### Deprecated
-
+-
 ### Removed
+-